backend/ai-service/requirements.txt
langgraph==0.2.53
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
langchain-groq==0.2.4
langchain-core==0.3.63
fastapi==0.115.5
uvicorn==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
duckduckgo-search==6.3.7
pydantic==2.10.3